requirements.txt
Django==1.6a1
South==0.8.1
djangorestframework==2.3.5

